Add a Cargo feature to select the field scalar implementation
This change adds a new "f128p" Cargo feature to the zk-circuit-builder
crate to allow selection of the field scalar implementation. The
"f128p" feature is not enabled by default but this change updates
witness-checker to enable the feature in its dependency configuration.
Consequently, this change should be functionality-preserving.

This change is intended to pave the way for adding scalar
implementations for other primes in the future. This change does not
add any other scalar implementations; it just wraps the existing scalar
implementation in the feature and then enables the feature in the build.

The functionality works by conditionally compiling the chosen scalar
implementation based on its Cargo feature name and then exposing the
resulting scalar struct under the "QuarkScalar" synonym that is used
throughout the codebase. If no scalar features are enabled, the build
will fail due to a missing QuarkScalar declaration. This change does not
provide any means to detect that multiple scalar features were enabled
since only one such feature exists at this time.