eval: don't call `SecretData::set` with results of lazy eval
It's hard to tell when we should actually call `s.set(..)` to cache the
results and when we shouldn't.  Calling it under the wrong conditions
causes a panic.  It's easiest to skip setting the value entirely;
eventually, each wire will be evaluated only once, so this additional
form of caching won't be needed.