main: wrap every `add_pass` function in a closure
This shouldn't be necessary, but it seems like the FnDef-to-FnPtr
coercion is picking `'static` as the lifetime argument for these
functions.  Then if a `CircuitFilter` that caches `Wire<'a>` internally
is added to the filter stack, rustc unifies `'a` with `'static` and
complains that "`arenas` must be borrowed for `'static`" (among other
errors).