Add `F128p` operations to the `PrivateOps` trait, and update the existing `PrivateOps` implementations.
The `PrivateOps` trait provides an interface for emitting private values. The `PrivateDirect` implementation
uses the `Sink` argument to achieve this by delegating to the `Sink::private_value{,_f128p}` methods. The
`Sink::private_value_f128p` method was introduced in bdac8fc. This commit extends `PrivateOps` with the necessary
methods for emitting `F128p` values. This includes directly emitting an `F128p` value, as well as private values that
are emitted in the process of compiling a `Switch` or `Cast` gate.

These new methods support the compilation of `ir::circuit::Wire`s whose type is `GF(Field::F128p)`.