Add support to `boolean::Backend` for compiling `ir::circuit::Wire` values that have `GF(Field::F128p)` type.
The `boolean::Backend` is responsible for compiling a circuit represented as a collection of `ir::circuit::Wire` values.
The primary way that users interact with `boolean::Backend` is through the `Backend` trait which exposes the `enforce_true`
method. The implementation of this method in `boolean::Backend` calls out to `Backend::convert_wires` which is the primary
function responsible for compiling wires. This supports compiling operations over `ir::circuit::Wire`s that have type
`GF(Field::F128p)` by matching on the type and delegating to the appropriate method(s) in the underlying `Sink`.

The changes to `boolean::Backend` enable the object to properly handle wires with `F128p` type. Previously,
the `bool::Backend` would have unpredictable behavior when given a circuit containing wires of `F128p` type.