Fix a bug in `lower::int_field_arith` where the lowering pass was not being applied recursively.
A subtraction gate (`a - b`) is implemented by lowering to an addition and negation gate over integers (`a + (-b)`), relying
on the compiler to recursively lower each of these operations to their `F128p` embeddings. This patch ensures that the
`IntFieldArith` lowering pass is recursively applied when necessary.

This bug was only identified as part of a manual integration test between the lowering pass and the `boolean::Backend` extended
with `F128p` support. The lowering of `a - b` to `a + (-b)` is _correct_ in the sense that evaluating the circuit gives the
correct answer. The bug is that the circuit is supposed to lower arithmetic operations exclusively to `F128p` operations. So,
this expectation was only relied on (and violated) when the circuit is consumed by `boolean::Backend` for production of SIEVE IR.