Refactor to ensure tests work as expected when the sieve_ir feature isn't enabled.
The Cargo sieve_ir feature is responsible for enabling support for emitting SIEVE IR
circuits in the backend. This commit adds cfg(feature = "sieve_ir") guards that are
necessary to ensure that the project builds, tests, etc. properly when the feature
isn't enabled.

In the process, the SieveIrField enum was moved to the back::boolean module from the
back::boolean::sink_sieve_ir module. The latter module is imported behind a sieve_ir
feature guard, which made importing SieveIrField unconditionally a compilation error.
Instead, having SieveIrField in back::boolean means it is available whether the sieve_ir
feature is enabled or not.