Reverse the bits of GateKind::Switch patterns for consistency with DietMC expectations.
The DietMC ZK interpreter expects integers to be represented with most significant digit
first. The current backend uses least significant digit first pervasively, and so this
commit contains a hack to reverse the bits of GateKind::Switch patterns. This is necessary so that
the GateKind::Switch condition and patterns will be equal when interpreted (incorrectly) as integers
with the most significant digit first.