Change the page size for private input values from GATE_PAGE_SIZE to PRIVATE_INPUT_PAGE_SIZE = 1024 * 1024.
The GATE_PAGE_SIZE is a limit on the number of gates in a SIEVE IR relation message, not the number of bytes.
It is intentionally set very low, because gates can be individually large (e.g. a Call gate with many
argument ranges). Instead, private input value messages have a small, fixed size and so should be given
a page size which is given in units of bytes and much larger than GATE_PAGE_SIZE. This helps with performance,
since the consumer libraries, deps/zkinterface-sieve{,-v3,-v5}, don't perform buffering.