Delete Bits::as_u64 and Bits::as_u8, as the former was unused and the latter only used in tests.
The Bits::as_u64 didn't properly account for possible trailing zeros, but it wasn't used anyway so
I just deleted it. The Bits::as_u8 made the same mistake, but was only used in tests/lower_int_field_arith.rs
and so I moved the function to that test module and deleted it from the public API of Bits.